+++ cpp/src/metrics/metrics.cu
@@ -20,6 +20,7 @@
 
 #include "metrics/entropy.h"
 #include "metrics/adjustedRandIndex.h"
+#include "metrics/mutualInfoScore.h"
 #include "metrics/randIndex.h"
 #include "score/scores.h"
 
@@ -56,6 +57,12 @@ double entropy(const cumlHandle &handle, const int *y, const int n, const int lo
   return MLCommon::Metrics::entropy(y, n, lower_class_range, upper_class_range, handle.getDeviceAllocator(), handle.getStream());
 
 
+double mutualInfoScore(const cumlHandle &handle, const int *y, const int *y_hat,
+                       const int n, const int lower_class_range,
+                       const int upper_class_range) {
+  return MLCommon::Metrics::mutualInfoScore(
+    y, y_hat, n, lower_class_range, upper_class_range,
+    handle.getDeviceAllocator(), handle.getStream());
 }
 
 }  // namespace Metrics